removed unused holiday program types (db)
moved holiday programs client to new naming conventions
