+++ client/src/components/holiday-programs/customer-booking-screen/components/step-2/cancellation-policy-modal.tsx
