+++ client/src/components/holiday-programs/customer-booking-screen/components/step-2/child-form.tsx
@@ -1,9 +1,9 @@
 import { DatePicker, Form, Input, Select } from 'antd'
 import dayjs from 'dayjs'
+import { AcuityConstants } from 'fizz-kidz'
 import React, { useState } from 'react'
 
 import { SimpleTextRule } from '@utils/formUtils'
-import { AcuityConstants } from 'fizz-kidz'
 
 const { TextArea } = Input
