+++ client/src/components/holiday-programs/customer-booking-screen/components/step-2/terms-and-conditions-modal.tsx
