+++ client/src/components/holiday-programs/customer-booking-screen/components/step-3/booking-summary.tsx
