+++ client/src/components/holiday-programs/holiday-program-class-selection-page.tsx
@@ -1,4 +1,5 @@
-import { AcuityConstants, AcuityTypes, ObjectEntries } from 'fizz-kidz'
+import type { AcuityTypes } from 'fizz-kidz'
+import { AcuityConstants, ObjectEntries } from 'fizz-kidz'
 import { DateTime } from 'luxon'
 import { useRef, useState } from 'react'
 import { useNavigate, useSearchParams } from 'react-router-dom'
@@ -7,7 +8,8 @@ import { Button, Paper } from '@mui/material'
 import { Skeleton } from '@mui/material'
 import FormControl from '@mui/material/FormControl'
 import MenuItem from '@mui/material/MenuItem'
-import Select, { SelectChangeEvent } from '@mui/material/Select'
+import type { SelectChangeEvent } from '@mui/material/Select'
+import Select from '@mui/material/Select'
 import Typography from '@mui/material/Typography'
 import { styled } from '@mui/material/styles'
 import { capitalise } from '@utils/stringUtilities'
