+++ server/fizz-kidz/src/discount-codes/index.ts
@@ -1,0 +1,9 @@
+export type DiscountCode = {
+    id: string
+    discountType: 'percentage' | 'price'
+    discountAmount: number
+    code: string
+    expiryDate: Date
+    numberOfUses: number
+    numberOfUsesAllocated: number
+}
