--- server/fizz-kidz/src/holidayPrograms/index.ts
@@ -1,46 +1,0 @@
-export interface BaseHolidayProgramBooking {
-    appointmentTypeId: number
-    dateTime: string
-    calendarId: number
-    parentFirstName: string
-    parentLastName: string
-    parentEmail: string
-    parentPhone: string
-    emergencyContactName: string
-    emergencyContactPhone: string
-    childName: string
-    // comes from client as birthday ISO string, saved in acuity as age.
-    // kept name as 'childAge' for backwards compat
-    childAge: string
-    childAllergies: string
-    childAdditionalInfo: string
-    discountCode: string
-    amountCharged: number
-    joinMailingList: boolean
-    title?: string // title of the program if it exists (coming from storyblok)
-    creations?: string[] // creations being made in the program if it exists (coming from storyblok)
-}
-
-interface BookedPaidHolidayProgram extends BaseHolidayProgramBooking {
-    booked: true
-    appointmentId: number
-}
-
-interface UnbookedPaidHolidayProgram extends BaseHolidayProgramBooking {
-    booked: false
-}
-
-export type PaidHolidayProgramBooking = BookedPaidHolidayProgram | UnbookedPaidHolidayProgram
-export type FreeHolidayProgramBooking = BaseHolidayProgramBooking
-
-export type HolidayProgramBooking = PaidHolidayProgramBooking | FreeHolidayProgramBooking
-
-export type DiscountCode = {
-    id: string
-    discountType: 'percentage' | 'price'
-    discountAmount: number
-    code: string
-    expiryDate: Date
-    numberOfUses: number
-    numberOfUsesAllocated: number
-}
